# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SE Project API (Team 8, May 2025)
  description: "Backend API for Team 8's Software Engineering Project (May 2025 Term)\n\
    \nThis backend powers a web-based mobile application designed for senior citizens\
    \ and their family caregivers. The application enhances daily living and care\
    \ coordination through these core modules identified based on user stories:\n\n\
    - Easy-to-use UI: Senior-friendly design with large text, bold colors, and simple\
    \ navigation\n- Medicine Tracker: Prescription management and medication identification\n\
    - Appointments Tracker: Book, view, and get reminders for doctor appointments\n\
    - SOS: Emergency alert system for designated contacts\n- Documents Vault: Secure\
    \ digital repository for medical reports and history\n\n---\n\n## User Story Mapping\n\
    \n| # | User Story (Summary) | API Endpoints |\n|---|----------------------|--------------|\n\
    | 1 | UI accessibility | (Frontend/UI, not API) |\n| 2 | Medicine management |\
    \ `/medicines/`, `/medicines/{medicine_id}`, `/medicines/transcribe` |\n| 3 |\
    \ Appointment management | `/appointments/`, `/appointments/user/{user_id}` |\n\
    | 4 | Emergency contact notification | `/users/{user_id}/sos/trigger`, `/emergency-contacts/`\
    \ |\n| 5 | Document storage | `/documents/`, `/documents/upload`, `/documents/user/{user_id}`\
    \ |\n| 6 | Medication adherence tracking | `/medicines/`, `/medicines/user/{user_id}`\
...
paths:
...
  /api/v1/documents/upload:
    post:
      tags:
      - Documents
      summary: Upload Document
      description: 'Upload a document file to S3 storage and return the file URL.


        Uses Amazon S3 to securely store medical documents and reports in the cloud,
        automatically generating unique URLs for each file and providing instant access
        to documents from anywhere.


        This endpoint accepts a file upload and stores it in S3, returning the public
        URL.

        The file will be given a unique name to prevent conflicts.


        Supported file types: PDF, DOC, DOCX, TXT, JPG, JPEG, PNG, GIF, CSV, XLS,
        XLSX

        Maximum file size: 10MB'
      operationId: upload_document_api_v1_documents_upload_post
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          type: integer
          title: User Id
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_documents_upload_post'
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Invalid file or upload error
        '401':
          description: Authentication required
        '413':
          description: File too large
        '500':
          description: Server error during upload
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/:
    post:
      tags:
      - Documents
      summary: Create Document
      description: 'Create a new document record for a user. Requires authentication.
        Clears cache for the user.


        Supports US5 by enabling users to store medical documents digitally.'
      operationId: create_document_api_v1_documents__post
      parameters:
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Session token for authentication
          title: Session Token
        description: Session token for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '201':
          description: Document created successfully.
        '400':
          description: Invalid input.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/user/{user_id}:
    get:
      tags:
      - Documents
      summary: Get Documents By User
      description: 'List all documents for a user. Results are cached for 5 minutes.


        Supports US5 and US9 by providing access to digital medical reports for users
        and doctors.'
      operationId: get_documents_by_user_api_v1_documents_user__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the user to get documents for
          title: User Id
        description: ID of the user to get documents for
      responses:
        '200':
          description: List of documents for the user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
                title: Response Get Documents By User Api V1 Documents User  User
                  Id  Get
        '404':
          description: User not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{document_id}:
    get:
      tags:
      - Documents
      summary: Get Document By Id
      description: 'Get a document by its ID.


        Supports US5 and US9 by letting users and doctors review specific medical
        documents.'
      operationId: get_document_by_id_api_v1_documents__document_id__get
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the document to retrieve
          title: Document Id
        description: ID of the document to retrieve
      responses:
        '200':
          description: Document found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Documents
      summary: Update Document
      description: 'Update an existing document record. Requires admin or owner. Clears
        the user''s cache.


        Supports US5 by keeping medical document records up to date.'
      operationId: update_document_api_v1_documents__document_id__put
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the document to update
          title: Document Id
        description: ID of the document to update
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdate'
      responses:
        '200':
          description: Document updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Documents
      summary: Delete Document
      description: 'Delete a document record by ID. Requires admin or owner. Clears
        the user''s cache.


        Supports US5 by allowing removal of outdated or incorrect medical documents.'
      operationId: delete_document_api_v1_documents__document_id__delete
      parameters:
      - name: document_id
        in: path
        required: true
        schema:
          type: integer
          description: ID of the document to delete
          title: Document Id
        description: ID of the document to delete
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Document deleted successfully.
          content:
            application/json:
              schema: {}
        '404':
          description: Document not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_document_api_v1_documents_upload_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_document_api_v1_documents_upload_post
...
    DocumentCreate:
      properties:
        name:
          type: string
          title: Name
          description: Name of the document
          example: Prescription.pdf
        file_url:
          type: string
          title: File Url
          description: URL to the uploaded document
          example: https://example.com/file.pdf
        user_id:
          type: integer
          title: User Id
          description: ID of the user who owns the document
          example: 1
      type: object
      required:
      - name
      - file_url
      - user_id
      title: DocumentCreate
      description: Schema for creating a new document record.
    DocumentResponse:
      properties:
        name:
          type: string
          title: Name
          description: Name of the document
          example: Prescription.pdf
        file_url:
          type: string
          title: File Url
          description: URL to the uploaded document
          example: https://example.com/file.pdf
        id:
          type: integer
          title: Id
          description: Unique document ID
          example: 1
        user_id:
          type: integer
          title: User Id
          description: ID of the user who owns the document
          example: 1
        timestamp:
          type: string
          format: date-time
          title: Timestamp
          description: Upload timestamp (ISO 8601)
          example: '2025-06-25T12:00:00+00:00'
      type: object
      required:
      - name
      - file_url
      - id
      - user_id
      - timestamp
      title: DocumentResponse
      description: Schema for returning document data to the client.
      example:
        file_url: https://example.com/file.pdf
        id: 1
        name: Prescription.pdf
        timestamp: '2025-06-25T12:00:00+00:00'
        user_id: 1
    DocumentUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Updated document name (optional)
          example: UpdatedName.pdf
        file_url:
          anyOf:
          - type: string
          - type: 'null'
          title: File Url
          description: Updated file URL (optional)
          example: https://example.com/updated.pdf
      type: object
      title: DocumentUpdate
      description: Schema for updating a document record (partial update).
    DocumentUploadResponse:
      properties:
        file_url:
          type: string
          title: File Url
          description: URL of the uploaded file
          example: https://example.com/file.pdf
        filename:
          type: string
          title: Filename
          description: Original filename
          example: document.pdf
        message:
          type: string
          title: Message
          description: Success message
          default: File uploaded successfully
      type: object
      required:
      - file_url
      - filename
      title: DocumentUploadResponse
      description: Schema for document upload response.
      example:
        file_url: https://example.com/file.pdf
        filename: document.pdf
        message: File uploaded successfully
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError